// src/frontend/src/app/api/fn/test-results/api-testresults-post-json.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { TestResultRequest } from '../../models/test-result-request';

export interface ApiTestresultsPost$Json$Params {
      body?: TestResultRequest
}

export function apiTestresultsPost$Json(http: HttpClient, rootUrl: string, params?: ApiTestresultsPost$Json$Params, context?: HttpContext): Observable<StrictHttpResponse<TestResultRequest>> {
  const rb = new RequestBuilder(rootUrl, apiTestresultsPost$Json.PATH, 'post');
  if (params) {
    rb.body(params.body, 'application/*+json');
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'text/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<TestResultRequest>;
    })
  );
}

apiTestresultsPost$Json.PATH = '/api/testresults';