// src/frontend/src/app/api/fn/test-results/api-grading-results-post-json.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { TestResultRequest } from '../../models/test-result-request';

export interface ApiGradingResultsPost$Json$Params {
      body?: TestResultRequest
}

export function apiGradingResultsPost$Json(http: HttpClient, rootUrl: string, params?: ApiGradingResultsPost$Json$Params, context?: HttpContext): Observable<StrictHttpResponse<TestResultRequest>> {
  const rb = new RequestBuilder(rootUrl, apiGradingResultsPost$Json.PATH, 'post');
  if (params) {
    rb.body(params.body, 'application/*+json');
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'text/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<TestResultRequest>;
    })
  );
}

apiGradingResultsPost$Json.PATH = '/api/grading-results';